Sync C# and VB implementations of AssemblyData subclasses
